requirements.txt
asgiref==3.8.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.13
django-decouple==2.1
django-health-check==3.17.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.7
openpyxl==3.1.3
pillow==10.3.0
psutil==5.9.5
psycopg2-binary==2.9.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
qrcode==8.2
reportlab==4.2.0
requests==2.31.0
six==1.17.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.4.0
